+++ CHANGES.rst
@@ -5,7 +5,7 @@ Change log for z3c.celery
 1.4.1 (unreleased)
 ==================
 
-- Nothing changed yet.
+- Annotate transaction with principal and task name, like zope.app.publication
 
 
 1.4.0 (2019-10-30)